fisrt version api rest with darto

+++ lib/controllers/todo_controller.dart
@@ -8,15 +8,22 @@ import '../database/database.dart';
 class TodoController {
   final db = Database.instance.db;
 
-  Future<void> getTodos(Request req, Response res) async {
-    final todos = db.select("SELECT * FROM todos");
+  void getTodos(Request req, Response res) {
+    final response = db.select("SELECT * FROM todos");
+    final todos = response.map((row) => Todo.fromMap(row)).toList();
+
     return res.json(todos);
   }
 
-  Future<void> getTodo(Request req, Response res) async {
+  void getTodo(Request req, Response res) {
     final id = req.params['id'];
+    final response = db.select("SELECT * FROM todos WHERE id = ?", [id]);
 
-    final todo = db.select("SELECT * FROM todos WHERE id = ?", [id]);
+    if (response.isEmpty) {
+      return res.status(NOT_FOUND).json({'error': 'Todo not exists.'});
+    }
+
+    final todo = response.map((row) => Todo.fromMap(row)).toList();
 
     return res.json(todo);
   }
@@ -26,17 +33,23 @@ class TodoController {
     final todo = Todo.fromMap(jsonDecode(body));
 
     db.execute(
-        'INSERT INTO todos (title, description, is_done) VALUES (?, ?, ?)',
-        [todo.title, todo.description, todo.isDone]);
+      'INSERT INTO todos (title, description, is_done) VALUES (?, ?, ?)',
+      [todo.title, todo.description, todo.isDone],
+    );
 
     return res.status(OK).end();
   }
 
-  Future<void> deleteTodo(Request req, Response res) async {
+  void deleteTodo(Request req, Response res) {
     final id = req.params['id'];
 
-    db.execute('DELETE FROM todos WHERE id = ?', [id]);
+    final response = db.select("SELECT * FROM todos WHERE id = ?", [id]);
 
+    if (response.isEmpty) {
+      return res.status(NOT_FOUND).json({'error': 'Todo not exists.'});
+    }
+
+    db.execute('DELETE FROM todos WHERE id = ?', [id]);
     return res.status(OK).end();
   }
 
@@ -54,8 +67,9 @@ class TodoController {
     final todo = Todo.fromMap(jsonDecode(body));
 
     db.execute(
-        'UPDATE todos SET title = ?, description = ?, is_done = ? WHERE id = ?',
-        [todo.title, todo.description, todo.isDone, id]);
+      'UPDATE todos SET title = ?, description = ?, is_done = ? WHERE id = ?',
+      [todo.title, todo.description, todo.isDone, id],
+    );
 
     return res.status(OK).end();
   }
